+++ test/e2e/tests/og-api/app/pages/api/og.js
@@ -2,7 +2,7 @@
 import { ImageResponse } from '@vercel/og'
 
 export const config = {
-  runtime: 'experimental-edge',
+  runtime: 'edge',
 }
 
 export default function () {
@@ -21,6 +21,6 @@ export default function () {
       >
         Hello!
       </div>
-    )
+    ),
   )
 }